refactor(effect): move account auth question modules
